requirements.txt
pyserial==3.5
rich==14.2.0
scipy==1.16.3
